# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0", 
      "title": "Dots Project CRUD",
      "description": "Dots Project Application API",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "http://dotsbackend-single-repository.apps.ocp.tmrnd.com.my",
    "basePath": "/",
    "tags": [
      {
        "name": "Users",
        "description": "Operations about user"
      },
      {
          "name": "Orders",
          "description": "Operations about order"
      }
      
    ],
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "components": {
        "securitySchemes" : {
            "bearerAuth" :{
                "type" : "http",
                "scheme" : "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
        "/users": {
            "get": {
                "tags": ["Users"],
                "summary": "Get all users in system",
                "security": [{
                    "bearerAuth": []
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Users"
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "tags" : ["Users"],
                "summary": "Get user profile in system",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT access token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "security": [{
                    "bearerAuth": []
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/users/login" :{
            "post": {
                "tags": ["Users"],
                "summary": "Logs user into the system",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "The user name for login",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/users/logout":{
            "post":{
                "tags": ["Users"],
                "summary": "Logs user out from system",
                "parameters": [
                    {
                        "name": "refreshToken",
                        "in": "body",
                        "description": "Valid refresh token",
                        "required": true,
                        "type": "object",
                        "properties": {
                            "refreshToken": {
                                "type": "string",
                                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI3YmNkMDk3ZWI0ZTRkYmYxMDQ0NzU0Yzk4YjBjYTY0YyIsImlhdCI6MTY1ODg2MDA3NiwiZXhwIjoxNjU5NDY0ODc2fQ.S69kF00FCM9aTGXi29-kRirLMGaeLWR15t7FI_GxtU4"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No response return"
                    },
                    "403": {
                        "description": "Invalid refresh token/refresh token expired"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "required": ["name", "username","staff_id","email","userlevel" ],
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "userlevel": {
                    "type": "integer"
                },
                "sektor": {
                    "type": "string"
                },
                "cost": {
                    "type": "string"
                },
                "mobile_no": {
                    "type": "string"
                },
                "del_no": {
                    "type": "string"
                },
                "staff_id": {
                    "type": "string"
                },
                "entgov": {
                    "type": "string"
                },
                "funct": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "Users": {
            "type": "array",
            "$ref": "#/definitions/User"
        },
        "LoginResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI3YmNkMDk3ZWI0ZTRkYmYxMDQ0NzU0Yzk4YjBjYTY0YyIsImlhdCI6MTY1ODg2MDA3NiwiZXhwIjoxNjU4ODYzNjc2fQ.Lf5ZBAfyFQNeyuLyEJ3bC9_8Z4f12qoKOiR0V02DrK0"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI3YmNkMDk3ZWI0ZTRkYmYxMDQ0NzU0Yzk4YjBjYTY0YyIsImlhdCI6MTY1ODg2MDA3NiwiZXhwIjoxNjU5NDY0ODc2fQ.S69kF00FCM9aTGXi29-kRirLMGaeLWR15t7FI_GxtU4"
                },
                "expiresAt": {
                    "type": "number",
                    "example": 1658863676
                }
            }
        }
    }
      
  }
  